updated configuation adapter
